Rush mvp

+++ client/src/Binary.cpp
@@ -49,11 +49,11 @@ namespace NmpBinary
             return NmpClient::Packet(event, x, y);
         } else if (event == NmpClient::EVENT::SPRITE) {
             NmpClient::SpriteInfo sprite = {
-                buffer[1],
-                buffer[2],
-                buffer[3],
-                buffer[4],
-                buffer[5]};
+                static_cast<int>(buffer[1]),
+                static_cast<int>(buffer[2]),
+                static_cast<int>(buffer[3]),
+                static_cast<int>(buffer[4]),
+                static_cast<int>(buffer[5])};
             return NmpClient::Packet(NmpClient::EVENT::SPRITE, sprite);
         } else if (event == NmpClient::EVENT::JOIN) {
             std::size_t id = static_cast<std::size_t>(buffer[1]);
